NowServing/requirements.txt
pandas==0.21.1
numpy==1.13.3
nltk==3.2.4
python-dotenv==0.7.1
scikit-learn==0.19.1
keras==2.1.3
tensorflow==1.6.0
embeddings

Flask==0.12.2
Flask-Cors==3.0.3
Flask-HTTPAuth==3.2.3
Flask-RESTful==0.3.6